Added folder "schemas" in which schemas for routes are separated by different files.
